improved delete function by adding a condition that cant delete if the out count is greater than 0

+++ Bogys-Winforms/Services/AdminFunctions/VideoLibraryFunctions.cs
@@ -71,22 +71,27 @@ namespace Bogys_Winforms.Services.AdminFunctions
                 video.VideoInCount = incount;
                 video.RentDays = rentdays;
                 video.VideoPrice = price;
-               
+
                 context.SaveChanges();
-                return true;                   
+                return true;
             }
         }
-
         public bool DeleteVideo(int videoId)
         {
             using (var context = new AppDbContext())
             {
                 var video = context.Video.FirstOrDefault(u => u.ID == videoId);
                 if (video == null) return false;
-            
+
+                if (video.VideoOutCount > 0)
+                {
+                    MessageBox.Show(strTxt.validateDelete, strTxt.validationTitle, MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                    return false;
+                }
+
                 context.Video.Remove(video);
                 context.SaveChanges();
-                return true;                    
+                return true;
             }
         }
